[refactor]: client can use custom view from ad data

+++ ad_sdk/src/main/java/com/example/ad_sdk/ad/AdData.kt
@@ -1,9 +1,7 @@
 package com.example.ad_sdk.ad
 
-import android.view.View
-
 data class AdData(
-    val id: Long,
+    val id: Int,
     val adName: String,
     val adImageUrl: String,
 )
